Remove the (minimal) support for bold on Windows
Firstly, it didn't actually *do* bold - it did something else, which in
some consoles (depending on the colour choices) might or might not
appear highlighted.

Secondly, the need to run a function on the text to get the effect
doesn't sit well with the 'page' function, which depends on the
control characters being embedded. The solution, of course, would
be to output all text through a function (instead of sys.stdout.write
and print) that on Windows detects the escape characters itself, and
does the appropriate thing.

This would be more complex, but given the first point, is probably not
worth the effort.

Also, conditionalise the import/usr of msvcrt so the program still works
when not on windows(!)